requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.1
PyPDF2>=3.0.1
langchain>=0.1.16
faiss-cpu>=1.7.4
huggingface-hub>=0.20.3
